+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.3] - 2026-08-06
+
+### Added
+
+- fetch-debian and fetch-ubuntu APIs for downloading Ubuntu/Debian patches.
+- list-debian and list-ubuntu to get specific ABI versions for downloading.
+
 ## [1.1.2] - 2026-08-06
 
 ### Added
@@ -54,8 +61,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created a base Python project for running tree-sitter queries on the Linux
   kernel source.
 
-[unreleased]: https://github.com/heavycircle/kconfig/compare/v1.2.0...HEAD
-[1.2.0]: https://github.com/heavycircle/kconfig/compare/v1.1.0...v1.2.0
+[unreleased]: https://github.com/heavycircle/kconfig/compare/v1.1.3...HEAD
+[1.1.3]: https://github.com/heavycircle/kconfig/compare/v1.1.2...v1.1.3
+[1.1.2]: https://github.com/heavycircle/kconfig/compare/v1.1.1...v1.1.2
+[1.1.1]: https://github.com/heavycircle/kconfig/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/heavycircle/kconfig/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/heavycircle/kconfig/compare/v0.1.0...v1.0.0
 [0.1.0]: https://github.com/heavycircle/kconfig/releases/tag/v0.1.0